requirements.txt
configparser
telethon
pymongo
google-generativeai
pillow
pdfplumber
python-magic
serpapi
requests
python-dotenv
googletrans==4.0.0-rc1